Make service code a common library, rename packages folder to packages_library
